增加注释, 修改部分逻辑

+++ terminal-server/src/main/java/com/qnxy/terminal/ProtocolVersion.java
@@ -7,20 +7,22 @@ import java.util.Arrays;
 import java.util.Optional;
 
 /**
+ * 协议版本
+ *
  * @author Qnxy
  */
 @RequiredArgsConstructor
 @Getter
 public enum ProtocolVersion {
 
-    VERSION_1(1),
+    VERSION_1((byte) 1),
 
     ;
 
 
-    private final int versionNumber;
+    private final byte versionNumber;
 
-    public static Optional<ProtocolVersion> valueOf(int versionNumber) {
+    public static Optional<ProtocolVersion> valueOf(byte versionNumber) {
         return Arrays.stream(values())
                 .filter(v -> v.getVersionNumber() == versionNumber)
                 .findFirst();
